moved bugreport into dialogs dir
