+++ src/components/dialogs/bugreports/BugReportWindow.stories.tsx
