+++ src/components/nav/MainMenu.tsx
@@ -2,14 +2,14 @@ import { Titlebar } from "@/components/nav/titlebar";
 import { WebviewWindow } from "@tauri-apps/api/webviewWindow";
 import { Outlet } from "react-router-dom";
 import { TopMenuItem } from "@/components/nav/TopMenuItem";
-import { FileDialog } from "../dialogs/FileDialog";
+import { FileDialog } from "../dialogs/file/FileDialog";
 import { useEffect, useState } from "react";
 import { invoke } from "@tauri-apps/api/core";
 import { ConnectorSettingsHandler } from "@/utils/connectorSettingsHandler";
 import { generateLibrary } from "@/library/utils/CustomWasmGenerator";
 import { UpdateWindow } from "../UpdateWindow";
 import { hasReadLatestPatchNotes } from "@/utils/UpdateChecker";
-import { BugReportWindow } from "../bugreports/BugReportWindow";
+import { BugReportWindow } from "../dialogs/bugreports/BugReportWindow";
 export const MainMenu: React.FC = () => {
   const connectorSettingsHandler = new ConnectorSettingsHandler();
   const [installWASMDialogOpen, setInstallWASMDialogOpen] =
